+++ longevity/src/test/scala/longevity/unit/model/annotations/PolyPersistentSpec.scala
@@ -24,9 +24,9 @@ class PolyPersistentSpec extends FlatSpec with GivenWhenThen with Matchers {
   }
 
   it should "create a companion object that extends PolyPType when there is no companion object" in {
-    PolyPNoCompanion.isInstanceOf[PolyPType[DomainModel, PolyPNoCompanion]] should be (true)
-    PolyPNoCompanion.asInstanceOf[PolyPType[DomainModel, PolyPNoCompanion]].pTypeKey should equal {
-      typeKey[PolyPNoCompanion]
+    Poly.isInstanceOf[PolyPType[DomainModel, Poly]] should be (true)
+    Poly.asInstanceOf[PolyPType[DomainModel, Poly]].pTypeKey should equal {
+      typeKey[Poly]
     }
   }