+++ longevity/src/test/scala/longevity/unit/model/annotations/testExamples/persistents.scala
@@ -28,15 +28,20 @@ case class PCaseClass()
 case class PCaseClassWithDefaults(x: Int = 7)
 
 @polyPersistent[DomainModel]
-trait PolyPNoCompanion
+sealed trait PolyPWithCompanion {
+  val z: Int
+}
+
+object PolyPWithCompanion {
+  index(props.z)
+  val y = 7
+}
+
+@derivedPersistent[DomainModel, PolyPWithCompanion]
+case class DerivedFromPolyPWithCompanion(z: Int) extends PolyPWithCompanion
 
 @polyPersistent[DomainModel]
-trait PolyPWithCompanion
-
-object PolyPWithCompanion { val y = 7 }
-
-@polyPersistent[DomainModel]
-trait Poly
+sealed trait Poly
 
 @derivedPersistent[DomainModel, Poly]
 class DerivedPNoCompanion extends Poly