+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [0.25.0] - 2016.07.XX - Small Fixes
 
+NOTES:
+- poly persistent and component now need to be sealed
+- PType now needs an explicit PEv. only matters if you are not using the annotations
+
 - 2016.07.12 - Add some specialized "implicit not found" error messages for implicit types in
   `longevity.model`.
 - 2016.07.12 - Reverse order of implicits for `Repo.retrieve` and `Repo.retrieveOne` methods. This